Fix incorrect use of X11 screen and output
There is a bug when xcalib is used with an X11 server configured in
ZaphodHead mode (where each physical display is a seperate X11 screen).
The second display cannot be selected with either:

	./xcalib -d :0.1 gamma_1_0.icc
	./xcalib -d :0 -s 1 gamma_1_0.icc

This is because a display's default screen is always selected and the
screen argument is incorrectly used to select the controller.

The patch uses the screen argument to correctly select and X11 screen
and adds an output argument to select which controller to use. This is
the same behaviour as the xrandr program.

Outputs can now be selected with:

	./xcalib -d :0.0 -o 1 gamma_1_0.icc
	./xcalib -d :0 -s 0 -o 1 gamma_1_0.icc

Screens (in ZaphodHead mode) can now be selected with:

	./xcalib -d :0.1 gamma_1_0.icc
	./xcalib -d :0 -s 1 gamma_1_0.icc